+++ chain/indexer/src/streamer/errors.rs
@@ -1,14 +1,15 @@
 use actix::MailboxError;
 
-/// Error occurs in case of failed data fetch
-#[derive(Debug)]
-pub enum FailedToFetchData {
-    MailboxError(MailboxError),
-    String(String),
-}
+use crate::streamer::CryptoHash;
 
-impl From<MailboxError> for FailedToFetchData {
-    fn from(actix_error: MailboxError) -> Self {
-        FailedToFetchData::MailboxError(actix_error)
-    }
+#[derive(thiserror::Error, Debug)]
+pub enum IndexerError {
+    #[error("Mailbox error: {0}")]
+    MailboxError(#[from] MailboxError),
+    #[error("Failed to fetch data: {0}")]
+    FailedToFetchData(String),
+    #[error("Failed to find local delayed receipt {receipt_id} in previous 1000 blocks")]
+    LocalDelayedReceiptNotFound { receipt_id: CryptoHash },
+    #[error("Internal error occured in indexer: {0}")]
+    InternalError(String),
 }